drivers: mmc: add mmc 8 bit cap quirk
For the Mars MA3 module it's possible to switch MMC hardware connection
between the 4 bit SD card and the 8 bit eMMC card. To handle this case
the MMC host has to support both 4 and 8 bit bus width what is
impossible in the current MMC driver implementation. This quirk allows
us to workaround this limitation.